Сorrection based on comments

+++ 1.2.01/main.cpp
@@ -4,7 +4,7 @@ int main()
 {
     std::setlocale(LC_ALL, "Russian");
 
-    double number;
+    int number;
 
     std::cout << "Введите число:\n";
     std::cin >> number;
